github-#23 - change type of variables holding movement delta to double for better precision.
Since I post movement events throught valuators it's possible to increase 'resolution' do doubles.

+++ include/gestures.h
@@ -58,7 +58,7 @@ struct Gestures {
 
 	/* Pointer movement is tracked here.
 	 */
-	int move_dx, move_dy;
+	double move_dx, move_dy;
 
 	/* Current time and time delta. Updated after each event and after sleeping.
 	 */
